Move TestDaemonPsud class to its own file; test API wrapper functions
