+++ sonic-psud/scripts/psud
@@ -93,7 +93,7 @@ def _wrapper_get_num_psus():
     return platform_psuutil.get_num_psus()
 
 
-def _wrapper_get_psus_presence(psu_index):
+def _wrapper_get_psu_presence(psu_index):
     if platform_chassis is not None:
         try:
             return platform_chassis.get_psu(psu_index - 1).get_presence()
@@ -102,7 +102,7 @@ def _wrapper_get_psus_presence(psu_index):
     return platform_psuutil.get_psu_presence(psu_index)
 
 
-def _wrapper_get_psus_status(psu_index):
+def _wrapper_get_psu_status(psu_index):
     if platform_chassis is not None:
         try:
             return platform_chassis.get_psu(psu_index - 1).get_powergood_status()
@@ -122,9 +122,9 @@ def get_psu_key(psu_index):
 def psu_db_update(psu_tbl, psu_num):
     for psu_index in range(1, psu_num + 1):
         fvs = swsscommon.FieldValuePairs([(PSU_INFO_PRESENCE_FIELD,
-                                           'true' if _wrapper_get_psus_presence(psu_index) else 'false'),
+                                           'true' if _wrapper_get_psu_presence(psu_index) else 'false'),
                                           (PSU_INFO_STATUS_FIELD,
-                                           'true' if _wrapper_get_psus_status(psu_index) else 'false')])
+                                           'true' if _wrapper_get_psu_status(psu_index) else 'false')])
         psu_tbl.set(get_psu_key(psu_index), fvs)
 
 
@@ -441,7 +441,7 @@ class DaemonPsud(daemon_base.DaemonBase):
 
     def _update_single_psu_data(self, index, psu, psu_tbl):
         name = get_psu_key(index)
-        presence = _wrapper_get_psus_presence(index)
+        presence = _wrapper_get_psu_presence(index)
         power_good = False
         voltage = None
         voltage_high_threshold = None
@@ -452,7 +452,7 @@ class DaemonPsud(daemon_base.DaemonBase):
         power = None
         is_replaceable = try_get(psu.is_replaceable, False)
         if presence:
-            power_good = _wrapper_get_psus_status(index)
+            power_good = _wrapper_get_psu_status(index)
             voltage = try_get(psu.get_voltage)
             voltage_high_threshold = try_get(psu.get_voltage_high_threshold)
             voltage_low_threshold = try_get(psu.get_voltage_low_threshold)
@@ -544,7 +544,7 @@ class DaemonPsud(daemon_base.DaemonBase):
         :return:
         """
         psu_name = get_psu_key(psu_index)
-        presence = _wrapper_get_psus_presence(psu_index)
+        presence = _wrapper_get_psu_presence(psu_index)
         fan_list = psu.get_all_fans()
         for index, fan in enumerate(fan_list):
             fan_name = try_get(fan.get_name, '{} FAN {}'.format(psu_name, index + 1))
